폴더 위치 변경, 트랜스폼 업데이트 적용

+++ SFML Framework/Framework/Scene.cpp
@@ -116,7 +116,7 @@ int Scene::FindGoAll(const std::string& name, std::list<GameObject*>& list)
 			list.push_back(obj);
 		}
 	}
-	return list.size();
+	return (int)list.size();
 }
 
 void Scene::ApplyAddGo()
